[update]  Support floating point number

+++ src/fr/redwoub/economysystem/commands/Balance.java
@@ -22,7 +22,7 @@ public class Balance implements CommandExecutor {
         Player player = (Player) sender;
 
         if(args.length == 0){
-            player.sendMessage(main.prefix + "§aVotre solde est de : " + file.getInt(player.getUniqueId().toString()) + "$");
+            player.sendMessage(main.prefix + "§aVotre solde est de : " + file.getDouble(player.getUniqueId().toString()) + "$");
         }else if(args.length == 1){
             Player target = Bukkit.getPlayer(args[0]);
             if(target == null){
@@ -30,7 +30,7 @@ public class Balance implements CommandExecutor {
                 return false;
             }
 
-            player.sendMessage("§aLe solde de : §f" + target.getName() + "§a est de : §e" + file.getInt(target.getUniqueId().toString()));
+            player.sendMessage("§aLe solde de : §f" + target.getName() + "§a est de : §e" + file.getDouble(target.getUniqueId().toString()));
         }else {
             player.sendMessage(main.prefix + "§cErreur §8: §e/balance (<player>)");
         }
